+++ src/domain/geometry/product.rs
@@ -1,7 +1,7 @@
 pub trait Product<Rhs = Self> {
     type Output;
 
-    fn dot(self, rhs: Rhs) -> f32;
+    fn dot(self, rhs: Rhs) -> f64;
 
     fn cross(self, rhs: Rhs) -> Self::Output;
 }
